Add application supervisor. Move init logic
