+++ lib/firmware_http.ex
@@ -1,5 +1,4 @@
 defmodule Nerves.Firmware.HTTP do
-
   @moduledoc """
   HTTP/JSON microservice to query and update firmware on a Nerves device.
 
@@ -31,13 +30,5 @@ defmodule Nerves.Firmware.HTTP do
                                       json_opts: [space: 1, indent: 2],
                                       timeout: 240_000
   """
-  @doc "Application start callback"
-  @spec start(atom, term) :: {:ok, pid} | {:error, String.t}
-  def start(_type, _args) do
-    port = Application.get_env(:nerves_firmware_http, :port, 8988)
-    path = Application.get_env(:nerves_firmware_http, :path, "/firmware")
-    timeout = Application.get_env(:nerves_firmware_http, :timeout, 120_000)
-    dispatch = :cowboy_router.compile [{:_,[{path, Nerves.Firmware.HTTP.Transport, []}]}]
-    :cowboy.start_http(__MODULE__, 10, [port: port], [env: [dispatch: dispatch], timeout: timeout])
-  end
+
 end
